+++ src/styles/colors.ts
@@ -1,4 +1,6 @@
 export enum Colors {
   black = '#000000',
   white = '#FFFFFF',
+  green = '#4FBA55',
+  grey = '#E5E5E5',
 }