+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.4 - 2020-12-07
+- Added component Button.
+- Added model TimerModel.
+- Added button for adding timers.
+- Added input to TimerDisplay.
+- Added fullWidth to Button.
+- Changed time to UTC.
+- Added layout Draggable.
+- Added drag and drop timers list.
+- Updated styles.
+
 ## 0.0.3 - 2020-12-05
 - Added component TimerClock.
 - Added component TimerDisplay.